--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashScreen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bed378b10f26f98418c2cbe76a122c25
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
