--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashScreenEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 110d2cd429722494ca93b8f61380b8d2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
