--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashScreenFadeOut.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c781e5102f1f124cb0e4f0d5984108b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
