--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashScreenManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b24b7fe9eedb994c8ffe95990a2d07c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
