--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashScreenManagerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8d13cb5b88cc85540ac1c2980c759614
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
