--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashScreenUIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b13ab4bbbb2737b48b97261ea664a74a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
