--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Core/SplashscreenDataManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb63debe028c190479ce0491aeac8835
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
