--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Data/DisclaimerData.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-namespace Studio23.SS2.SplashScreenSystem.Data
-{
-    [System.Serializable]
-    public class DisclaimerData : ScriptableObject
-    {
-        public string DisclaimerTitle;
-        public string DisclaimerDescription;
-    }
-}
-
