+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Data/ScreenTextData.cs.meta
