+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Data/SplashScreenData.cs
@@ -1,0 +1,13 @@
+using System;
+using UnityEngine;
+
+namespace Studio23.SS2.SplashScreenSystem.Data
+{
+    [Serializable]
+    public class SplashScreenData
+    {
+        public float duration;
+        public float fadeduration;
+        public ScriptableObject data;
+    }
+}
