+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/Data/SplashScreenData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0d321961f6d9f44439b2a3affc0cd002
+guid: 2038d99f7b9ca66409f1232fb51f7a84
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
