--- Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/FinishHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9ff523ad18f553449665fa0f39fcda7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
