+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/UI.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3dd6b31e99826674b8d05eb8180fd7b1
+guid: 94e8873bfd1a3ad4eae22b03e19e12cf
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
