+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Runtime/UI/UIManager.cs.meta
