+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Data.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 40801e1678084064d86f23fc1fd12f3c
+guid: a34912b306284704aa692462308cebb0
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
