+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Data/Splash Screen Data/EULA.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a7a862d03998aa14995427ca268090eb, type: 3}
+  m_Name: EULA
+  m_EditorClassIdentifier: 
+  Title: EULA
+  Description: EULA
+  ShowButton: 1
