+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Data/Splash Screen Data/EULA.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 108795437e0608a4f8ca9b21c8061e5c
-folderAsset: yes
-DefaultImporter:
+guid: fdd5e5a611325c749b9420b1eeff4552
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
