+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Data/Splash Screen Data/ThirdParty.asset.meta
