+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Prefabs.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 9fc0d4010bbf28b4594072e72b8655ab
+guid: 96738a08b4859ad4c94b699912ee9494
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
