+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Prefabs/ThirdPartyImagePrefab.prefab.meta
