+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/SampleScene.unity.meta
