+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Sprite.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: b8520c1d264ae1240819ba1cfbb15d5e
+guid: ca27f0ceb1354ca418484513b04f805c
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
