+++ Assets/Packages/com.studio23.ss2.splashscreensystem/Samples/Sprite/fmod_logo.png
