--- Assets/Resources/SceneTable.cs
@@ -1,7 +1,0 @@
-namespace Studio23.SS2.SceneLoadingSystem.Data
-{
-	public static class SceneTable
-	{
-		public static readonly string SplashScreen = "SplashScreen";
-	}
-}
