--- Assets/Resources/SceneTable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8279361cdbd77b54eb151ffc1ea01995
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
