--- Assets/Resources/Splash Screen Data/EULA.asset
@@ -1,44 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0d321961f6d9f44439b2a3affc0cd002, type: 3}
-  m_Name: EULA
-  m_EditorClassIdentifier: 
-  EulaTitle: EULA
-  EulaDescription: Contrary to popular belief, Lorem Ipsum is not simply random text.
-    It has roots in a piece of classical Latin literature from 45 BC, making it over
-    2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College
-    in Virginia, looked up one of the more obscure Latin words, consectetur, from
-    a Lorem Ipsum passage, and going through the cites of the word in classical literature,
-    discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and
-    1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil) by
-    Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very
-    popular during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum dolor
-    sit amet..", comes from a line in section 1.10.32. Contrary to popular belief,
-    Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin
-    literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin
-    professor at Hampden-Sydney College in Virginia, looked up one of the more obscure
-    Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites
-    of the word in classical literature, discovered the undoubtable source. Lorem
-    Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus Bonorum et Malorum"
-    (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise
-    on the theory of ethics, very popular during the Renaissance. The first line
-    of Lorem Ipsum, "Lorem ipsum dolor sit amet..", comes from a line in section
-    1.10.32. Contrary to popular belief, Lorem Ipsum is not simply random text. It
-    has roots in a piece of classical Latin literature from 45 BC, making it over
-    2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College
-    in Virginia, looked up one of the more obscure Latin words, consectetur, from
-    a Lorem Ipsum passage, and going through the cites of the word in classical literature,
-    discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and
-    1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil) by
-    Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very
-    popular during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum dolor
-    sit amet..", comes from a line in section 1.10.32.
