--- Assets/Resources/Splash Screen Data/EULA.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f737e2d2cac428c4ca3d93c8a90c97d7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
