--- Assets/dummy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 91ce2cd8b228f424e9bb741f43b25acc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
